+++ R/ddbs_predicates.R
@@ -225,11 +225,17 @@ ddbs_predicate <- function(
   
  # 4. Build predicate expression
   if (st_predicate == "ST_DWithin") {
+
+    ## Warn if the distance arg wasn't specified, and give it a value of 0
+    if (is.null(distance)) {
+      cli::cli_warn("{.val distance} wasn't specified. Using ST_Within.")
+      distance <- 0
+    }
     
     ## check the CRS units to use the right function
     crs_units <- crs_x$units_gdal
     if (crs_units != "metre") {
-      predicate_expr <- glue::glue("ST_DWithin_Spheroid(ST_FlipCoordinates(x.{x_geom}), ST_FlipCoordinates(y.{y_geom}), {distance})")
+      predicate_expr <- glue::glue("ST_DWithin_Spheroid(x.{x_geom}, y.{y_geom}, {distance})")
       if (crs_x$input != "EPSG:4326") {
         cli::cli_warn(
           "Inputs are in {.val {crs_x$input}}, not {.val EPSG:4326}. Distance calculations may be less accurate. Consider transforming to {.val EPSG:4326} or a projected CRS."
@@ -239,10 +245,6 @@ ddbs_predicate <- function(
       predicate_expr <- glue::glue("ST_DWithin(x.{x_geom}, y.{y_geom}, {distance})")
     }
     
-    if (is.null(distance)) {
-      cli::cli_warn("{.val distance} wasn't specified. Using ST_Within.")
-      distance <- 0
-    }
     
   } else {
     predicate_expr <- glue::glue("{st_predicate}(x.{x_geom}, y.{y_geom})")