+++ R/ddbs_templates.R
@@ -324,7 +324,7 @@ template_measure <- function(
   if (crs_units == "metre") {
       st_function <- glue::glue("{fun}({x_geom})")
   } else {
-      st_function <- glue::glue("{fun}_Spheroid(ST_FlipCoordinates({x_geom}))")
+      st_function <- glue::glue("{fun}_Spheroid({x_geom})")
   }
   
   ## 3.4. Determine units for output